requirements.txt
# CPath-Omni Dependencies
# Pinned versions for reproducibility

# Core ML frameworks
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78
matplotlib==3.8.2
seaborn==0.13.0

# Progress bars
tqdm==4.66.1

# Vision models
open-clip-torch==2.23.0
timm==0.9.12

# Text models / Transformers
transformers==4.36.0
accelerate==0.25.0
tokenizers==0.15.0
safetensors==0.4.1

# Whole slide image processing
openslide-python==1.3.1

# Grad-CAM visualization
grad-cam==1.4.8

# Geometry (for annotations)
shapely==2.0.2

# Configuration
pyyaml==6.0.1
omegaconf==2.3.0

# Jupyter notebooks (optional)
jupyter==1.0.0
ipywidgets==8.1.1
ipykernel==6.27.1

# Development (optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Configuration
pyyaml>=6.0.0
omegaconf>=2.3.0
